Moved types around
