+++ src/pages/PlaygroundPage.svelte
@@ -20,7 +20,7 @@
   import PlaygroundLog from '../components/playground/PlaygroundLog.svelte';
   import MicrobitAccelerometerDataSynthesizer from '../components/playground/inputSynthesizer/MicrobitAccelerometerDataSynthesizer.svelte';
   import LiveDataBufferUtilizationPercentage from '../components/playground/LiveDataBufferUtilizationPercentage.svelte';
-    import StaticConfiguration from '../StaticConfiguration';
+  import StaticConfiguration from '../StaticConfiguration';
 
   const getRandomGesture = (): Gesture => {
     return gestures.getGestures()[
@@ -32,9 +32,7 @@
   const trainModelButtonClicked = () => {
     playgroundContext.addMessage('training model...');
     model
-      .train(
-        new LayersModelTrainer(StaticConfiguration.layersModelTrainingSettings),
-      )
+      .train(new LayersModelTrainer(StaticConfiguration.layersModelTrainingSettings))
       .then(() => {
         playgroundContext.addMessage('Finished training!');
       });
