+++ src/pages/model/tileview/ModelPageTileView.svelte
@@ -16,7 +16,7 @@
   import Microbits from '../../../script/microbit-interfacing/Microbits';
   import ModelPageTileViewTiles from './ModelPageTileViewTiles.svelte';
   import { classifier } from '../../../script/stores/Stores';
-    import StaticConfiguration from '../../../StaticConfiguration';
+  import StaticConfiguration from '../../../StaticConfiguration';
 
   // In case of manual classification, variables for evaluation
   let recordingTime = 0;
@@ -66,7 +66,7 @@
       return;
     }
 
-    let shouldClassify: boolean = (buttons.buttonA === 1 || buttons.buttonB === 1);
+    let shouldClassify: boolean = buttons.buttonA === 1 || buttons.buttonB === 1;
 
     if (shouldClassify) {
       classifyClicked();
